field renames and bugfixes
renamed all fields. They now either start with a_ (arithmetic), v_ (value) or c31_ (cvss vector content). This is to avoid field collisions in cases where for example a field "confidentiality" already exists in the output of the search. The macro would "overwrite" this existing field. The field "confidentiality" is now named "c31_confidentialtity" - that also doesn't rule out a collision but it makes it more unlikely.

renaming the fields allowed me to remove all temporary fields by adding this line:
> | fields - a_*, v_*, c31_*
that makes debugging of full reports a bit easier, since the event is not flooded with hundreds of uninteresting fields from the cvss calculation

fixed a bug which caused a "modified attack vector" of "H" to trigger an error even though its a perfectly fine value